+++ NEWS.md
@@ -1,4 +1,6 @@
 ## v2.0.0 (IN PROGRESS)
+- Add LIGHT_RESOURCE type [MODLD-980](https://folio-org.atlassian.net/browse/MODLD-980)
+- Concept label generation now orders labels as FOCUS first, followed by sorted SUB_FOCUS entries [MODLD-913](https://folio-org.atlassian.net/browse/MODLD-913)
 
 ## v1.9.9 (13-02-2026)
 - Update Mapping for 300$b & $e (Physical description & Accompanying material) [MODLD-508](https://folio-org.atlassian.net/browse/MODLD-508)
@@ -30,7 +32,6 @@
 - Update the IRI of ID_LOCAL to http://library.link/identifier/LOCAL [MODLD-939](https://folio-org.atlassian.net/browse/MODLD-939)
 - Add resource types FAST, GND, LCCSH, LCDGT, LCGFT, LCMPT, LCNAF, LCSH, MESH, VIAF and WIKIID [MODLD-939](https://folio-org.atlassian.net/browse/MODLD-939)
 - Create HubLabelGenerator to generate labels for Hubs [MODLD-970](https://folio-org.atlassian.net/browse/MODLD-970)
-- Add LIGHT_RESOURCE type [MODLD-980](https://folio-org.atlassian.net/browse/MODLD-980)
 
 ## v1.0.1 (03-12-2025)
 - Initial release